pyproject.toml
[project]
name = "va-decision-test"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "psycopg[binary]>=3.1",
    "pgvector>=0.3",
    "google-genai>=1.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "textstat>=0.7",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]
api = ["fastapi>=0.115", "uvicorn[standard]>=0.32"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
